requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.7.0
dbt-common==1.11.0
dbt-core==1.9.0b2
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
idna==3.10
importlib-metadata==6.11.0
importlib-resources==6.4.5
isodate==0.6.1
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
leather==0.4.0
MarkupSafe==2.1.5
mashumaro==3.14
more-itertools==10.5.0
msgpack==1.1.0
networkx==3.1
numpy==1.24.4
ordered-set==4.1.0
packaging==24.2
pandas==2.0.3
parsedatetime==2.6
pathspec==0.12.1
pkgutil-resolve-name==1.3.10
protobuf==4.25.6
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.3
text-unidecode==1.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
zipp==3.20.2
